Introduce custom signed int type that is compatiple with rlp non-support for negative int encoding

+++ state/validation.go
@@ -18,8 +18,8 @@ func validateBlock(state LastestBlockState, block *types.Block) error {
 	//if block.Header().ChainID != state.ChainID {
 	//	return fmt.Errorf("Wrong Block.Header.ChainID. Expected %v, got %v", state.ChainID, block.ChainID)
 	//}
-	if block.Header().Height != state.LastBlockHeight.Uint64()+1 {
-		return fmt.Errorf("Wrong Block.Header.Height. Expected %v, got %v", state.LastBlockHeight.Uint64()+1, block.Height)
+	if int64(block.Header().Height) != state.LastBlockHeight.Int64()+1 {
+		return fmt.Errorf("Wrong Block.Header.Height. Expected %v, got %v", state.LastBlockHeight.Int64()+1, block.Height)
 	}
 	/*	TODO: Determine bounds for Time
 		See blockchain/reactor "stopSyncingDurationMinutes"
